fix: header router updated
