+++ package.json
@@ -9,8 +9,10 @@
     "feather-icons-react": "^0.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-markdown": "^7.1.0",
     "react-modal": "^3.14.3",
-    "react-router-dom": "^5.3.0",
+    "react-router": "^6.0.2",
+    "react-router-dom": "6.0.2",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.0.1"
   },
